% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_match_score.R
\name{comp.match.acc}
\alias{comp.match.acc}
\title{Compute match accuracies.}
\usage{
comp.match.acc(mat)
}
\arguments{
\item{mat}{A match score matrix}
}
\value{
A data frame containing accuracies under four matching schemes.
\code{one_to_one}, \code{SNPquery}, \code{STRquery}, \code{needle_in_haystack}.
}
\description{
Compute match accuracies given a match score matrix under four schemes:
one-to-one (Hungarian algorithm), one-to-many:pick STR, one-to-many: pick SNP, one match.
}
\details{

}
