% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_match_score.R
\name{comp.match.mat}
\alias{comp.match.mat}
\title{Compute match score matrix based on a single STR locus.}
\usage{
comp.match.mat(cott.vec, marker, str.f)
}
\arguments{
\item{cott.vec}{A vector of Cotterman identity coefficients representing
the test relatedness hypothesis.}

\item{marker}{The name of a STR locus under consideration.}

\item{str.f}{A name (including full path) of a file containing
\emph{known (NOT imputed)} STR genotypes at the STR locus \code{marker}.}
}
\value{
A match score matrix based on a single STR locus. An \verb{(i,j)} entry of
the match score matrix is a match score between individual \code{i} with a STR \emph{observed} genotype
and individual \code{j} with an \emph{observed} SNP profile based on a single STR locus \code{marker}
and a test relatedness hypothesis \code{cott.vec}.
}
\description{
insert description.
}
\note{
\code{str.f} should be a \code{GT} file format of \code{vcftools}. The first column should be
a chromosome and the second column should be a marker genomic position.
The sample data starts from the third column. The first row should be a header
and the second row should be data. The genotype can be either a phased or unphased format.
}
