% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imput_str.R
\name{impute.str}
\alias{impute.str}
\title{Impute STR genotypes from surrounding SNPs and extract imputed STR genotypes and genotype probabilites.}
\usage{
impute.str(
  snp.f,
  ref.f,
  map.f,
  marker,
  nthreads = 1,
  niterations = 10,
  maxlr = 1e+06,
  lowmem = "false",
  window = 50000,
  overlap = 3000,
  cluster = 0.005,
  ne = 1e+06,
  err = 1e-04,
  seed = -99999,
  modelscale = 0.8
)
}
\arguments{
\item{snp.f}{A name (including full path) of a file containing SNPs.
Must be a standard vcf format.}

\item{ref.f}{A name (including full path) of a file containing SNP-STR
haplotypes to be used as a reference panel for imputation.
\emph{Note that the reference panel must be phased and have no
missing values.}}

\item{map.f}{A name (including full path) of a file genetic map in plink format.}

\item{marker}{Name of a STR locus to be imputed.}

\item{nthreads}{(positive integer) Number of threads for BEAGLE, default=1.}

\item{niterations}{(positive integer) Number of phasing iterations in BEAGLE, default=10.}

\item{maxlr}{(number ≥ 1) The maximum likelihood ratio at a genotype, default=1000000.}

\item{lowmem}{(true/false) Whether a memory efficient algorithm should be used, default=false.}

\item{window}{(positive integer) The number of markers to include in each sliding window, default=50000.}

\item{overlap}{(positive integer) The number of markers of overlap between sliding windows, default=3000.}

\item{cluster}{(non-negative number) The maximum cM distance between individual markers that are
combined into an aggregate marker when imputing ungenotyped markers, default=0.005.}

\item{ne}{(integer) The effective population size when imputing ungenotyped markers, default=1000000.}

\item{err}{(nonnegative number) The allele miscall rate, default=0.0001.}

\item{seed}{(integer) The seed for the random number generator, default=-99999.}

\item{modelscale}{(positive number) the model scale parameter when sampling haplotypes
for unrelated individuals, default=0.8.}
}
\value{
None
}
\description{
This function imputes genotypes of individuals at a STR \code{marker} locus from its
surrounding SNPs \code{snp.f} using a reference panel \code{ref.f} and a genetic map \code{map.f}.
}
\details{
This function imputes genotypes of individuals at a STR \code{marker} locus from its
surrounding SNPs \code{snp.f} using a reference panel \code{ref.f} and a genetic map \code{map.f}.
\emph{The reference panel must be phased and have no missing values.}
The imputed and phased SNP-STR vcf file is stored at \code{(base.dir)/imputed_str} where
the \code{base.dir} was set by running \code{\link{setup}} in the beginning of the pipeline.
From the imputed SNP-STR vcf, the function extracts imputed genotypes and genotype probabilities
at the STR \code{marker} locus and save it in the \code{(base.dir)/imputed_str} folder.
The output file name containing imputed genotypes probabilities is \verb{imp_str_(marker).GP.FORMAT}.
}
\seealso{
\href{https://faculty.washington.edu/browning/beagle/beagle_4.1_21Jan17.pdf}{BEAGLE manual}.
\href{https://github.com/vcftools/vcftools}{vcftools manual}.
}
