% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_ref.R
\name{phase.ref}
\alias{phase.ref}
\title{Phase reference vcf file.}
\usage{
phase.ref(
  ref.f,
  nthreads = 1,
  niterations = 10,
  lowmem = "false",
  window = 50000,
  overlap = 3000,
  impute = "true",
  cluster = 0.005,
  gprobs = "false",
  ne = 1e+06,
  err = 1e-04,
  seed = -99999,
  modelscale = 0.8
)
}
\arguments{
\item{ref.f}{A name (including full path) of a file containing SNP-STR
haplotypes to be used as a reference panel for imputation.
Note that the reference panel must be phased and have no
missing values.}

\item{nthreads}{(positive integer) Number of threads for BEAGLE, default=1.}

\item{niterations}{(positive integer) Number of phasing iterations in BEAGLE, default=10.}

\item{lowmem}{(true/false) Whether a memory efficient algorithm should be used, default=false.}

\item{window}{(positive integer) The number of markers to include in each sliding window, default=50000.}

\item{overlap}{(positive integer) The number of markers of overlap between sliding windows, default=3000.}

\item{cluster}{(non-negative number) The maximum cM distance between individual markers that are
combined into an aggregate marker when imputing ungenotyped markers, default=0.005.}

\item{ne}{(integer) The effective population size when imputing ungenotyped markers, default=1000000.}

\item{err}{(nonnegative number) The allele miscall rate, default=0.0001.}

\item{seed}{(integer) The seed for the random number generator, default=-99999.}

\item{modelscale}{(positive number) the model scale parameter when sampling haplotypes
for unrelated individuals, default=0.8.}

\item{snp.f}{A name (including full path) of a file containing SNPs.
Must be a standard vcf format.}

\item{map.f}{A name (including full path) of a file genetic map in plink format.}

\item{marker}{Name of a STR locus to be imputed.}

\item{maxlr}{(number ≥ 1) The maximum likelihood ratio at a genotype, default=1000000.}
}
\value{
None
}
\description{
This function phase reference vcf file (\code{ref.f}) with \code{BEAGLE}.
}
\details{
The phased reference file is to be used as a reference panel for \code{BEAGLE} imputation.
The output file is saved under \code{ref_phased} folder within the base folder.
The output file name is the same as \code{ref.f} but with \verb{_phs} postfix.
}
\seealso{
\href{https://faculty.washington.edu/browning/beagle/beagle_4.1_21Jan17.pdf}{BEAGLE manual}.
}
