% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_ref.R
\name{ref.al.freq}
\alias{ref.al.freq}
\title{Compute reference STR allele frequencies from a vcf file.}
\usage{
ref.al.freq(ref.f, marker)
}
\arguments{
\item{ref.f}{A name (including full path) of a file containing SNP-STR
haplotypes to be used as a reference panel for imputation.
Note that the reference panel must be phased and have no
missing values.}

\item{marker}{STR marker name of which allele frequency is computed.}
}
\value{
None
}
\description{
This function extract STR allele frequences from reference vcf file (\code{ref.f}) with \code{vcftools}.
}
\details{
The STR allele frequency is necessary for computing match scores.
The output file is saved under \code{ref_alfrq} folder within the base folder.
The output file name is \verb{ref_(marker).frq} format.
}
\seealso{
\href{https://github.com/vcftools/vcftools}{vcftools manual}.
}
