% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Setup environmental variables.}
\usage{
setup(base.dir, bgl.jar, vcf.exe)
}
\arguments{
\item{base.dir}{Full path of the base folder of the pipeline.
All the output of the pipeline will be saved here.
If the folder doesn't exist, it's automatically created.}

\item{bgl.jar}{Full path to beagle executable.
Format must be: (full path to dir)/beagle.(version).jar.
version is the Beagle version code (eg. “01Oct15.6a3”).}

\item{vcf.exe}{Full path to vcftools executable.}
}
\value{
None
}
\description{
This function is meant to be run at the very beginning of the pipeline to set
package environmental variables.
}
